@@ -3,13 +3,16 @@ package com.atilsamancioglu.kotlincountdown
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.os.CountDownTimer
-import kotlinx.android.synthetic.main.activity_main.*
+import com.atilsamancioglu.kotlincountdown.databinding.ActivityMainBinding
 
 class MainActivity : AppCompatActivity() {
+    private lateinit var binding: ActivityMainBinding
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
+        binding = ActivityMainBinding.inflate(layoutInflater)
+        val view = binding.root
+        setContentView(view)
 
 
         //Abstract - Inheritance
@@ -17,11 +20,11 @@ class MainActivity : AppCompatActivity() {
 
         object : CountDownTimer(10000,1000) {
             override fun onFinish() {
-                textView.text = "Left: 0"
+                binding.textView.text = "Left: 0"
             }
 
             override fun onTick(millisUntilFinished: Long) {
-                textView.text = "Left: ${millisUntilFinished/1000}"
+                binding.textView.text = "Left: ${millisUntilFinished/1000}"
             }
 
         }.start()
